+++ ThirdPartyAdapters/chartboost/chartboost/src/main/java/com/google/ads/mediation/chartboost/ChartboostAdapterUtils.java
@@ -99,6 +99,7 @@ class ChartboostAdapterUtils {
               + "in the AdMob or Ad Manager UI.");
       return false;
     }
+
     return true;
   }