+++ ThirdPartyAdapters/mintegral/mintegral/src/main/java/com/google/ads/mediation/mintegral/MintegralConstants.java
@@ -2,22 +2,29 @@ package com.google.ads.mediation.mintegral;
 
 import androidx.annotation.IntDef;
 import androidx.annotation.NonNull;
+
 import com.google.android.gms.ads.AdError;
+
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 
 public class MintegralConstants {
 
-  public static final String AD_UNIT_ID = "ad_unit_id";
-  public static final String PLACEMENT_ID = "placement_id";
   public static final String APP_ID = "app_id";
   public static final String APP_KEY = "app_key";
+  public static final String AD_UNIT_ID = "ad_unit_id";
+  public static final String PLACEMENT_ID = "placement_id";
   public static final String ERROR_DOMAIN = "com.google.ads.mediation.mintegral";
   public static final String MINTEGRAL_SDK_ERROR_DOMAIN = "com.mbridge.msdk";
 
   @Retention(RetentionPolicy.SOURCE)
-  @IntDef(value = {ERROR_MINTEGRAL_SDK,
-      ERROR_INVALID_SERVER_PARAMETERS})
+  @IntDef(
+      value = {ERROR_INVALID_SERVER_PARAMETERS,
+          ERROR_BANNER_SIZE_UNSUPPORTED,
+          ERROR_INVALID_BID_RESPONSE,
+          ERROR_MINTEGRAL_SDK,
+          ERROR_CODE_NO_FILL,
+      })
   public @interface AdapterError {
 
   }
@@ -28,10 +35,24 @@ public class MintegralConstants {
   public static final int ERROR_MINTEGRAL_SDK = 100;
 
   /**
-   * Invalid server parameters (e.g Missing app ID or placement ID).
+   * Invalid server parameters (e.g. Missing App ID or Placement ID).
    */
   public static final int ERROR_INVALID_SERVER_PARAMETERS = 101;
 
+  /**
+   * The requested ad size does not match a Mintegral supported banner size.
+   */
+  public static final int ERROR_BANNER_SIZE_UNSUPPORTED = 102;
+
+  /**
+   * Missing or invalid bid response.
+   */
+  public static final int ERROR_INVALID_BID_RESPONSE = 103;
+
+  /**
+   * Mintegral SDK returned a no fill error.
+   */
+  public static final int ERROR_CODE_NO_FILL = 104;
 
   @NonNull
   public static AdError createAdapterError(@AdapterError int errorCode,