+++ ThirdPartyAdapters/applovin/CHANGELOG.md
@@ -1,5 +1,19 @@
 ## AppLovin Android Mediation Adapter Changelog
 
+#### Version 11.6.0.0
+- Verified compatibility with AppLovin SDK 11.6.0.
+
+Built and tested with:
+- Google Mobile Ads SDK version 21.3.0.
+- AppLovin SDK version 11.6.0.
+
+#### Version 11.5.5.0
+- Verified compatibility with AppLovin SDK 11.5.5.
+
+Built and tested with:
+- Google Mobile Ads SDK version 21.3.0.
+- AppLovin SDK version 11.5.5.
+
 #### Version 11.5.4.0
 - Verified compatibility with AppLovin SDK 11.5.4.