+++ ThirdPartyAdapters/ironsource/CHANGELOG.md
@@ -1,5 +1,12 @@
 ## IronSource Android Mediation Adapter Changelog
 
+#### Version 7.2.6.0
+- Verified compatibility with ironSource SDK version 7.2.6.
+
+Built and tested with:
+- Google Mobile Ads SDK version 21.3.0.
+- IronSource SDK version 7.2.6.
+
 #### Version 7.2.5.0
 - Verified compatibility with ironSource SDK version 7.2.5.
 - Updated the minimum required Google Mobile Ads SDK version to 21.3.0.