+++ ThirdPartyAdapters/mintegral/CHANGELOG.md
@@ -1,8 +1,19 @@
 ## Mintegral Android Mediation Adapter Changelog
 
-#### Initial Version
-- Initial release!
+#### Next Version
+- Updated the adapter to forward `onUserEarnedReward()` before `onAdClosed()` when showing rewarded ads.
+
+#### Version 16.3.11.0 (In progress)
+- Verified compatibility with Mintegral SDK 16.3.11.
 
 Built and tested with:
-- Google Mobile Ads SDK version 21.1.0.
-- Mintegral SDK version 16.2.11.
+- Google Mobile Ads SDK version 21.3.0.
+- Mintegral SDK version 16.3.11.
+
+#### Version 16.2.61.0
+- Initial release!
+- Added bidding support for banner (includes MREC), interstitial, rewarded and native ad formats.
+
+Built and tested with:
+- Google Mobile Ads SDK version 21.3.0.
+- Mintegral SDK version 16.2.61.