+++ ThirdPartyAdapters/mytarget/CHANGELOG.md
@@ -1,5 +1,27 @@
 ## myTarget Android Mediation Adapter Changelog
 
+#### Version 5.16.3.0 (In progress)
+- Verified compatibility with myTarget SDK version 5.16.3.
+- Updated the minimum required Google Mobile Ads SDK version to 21.4.0.
+
+Built and tested with:
+- Google Mobile Ads SDK version 21.4.0.
+- MyTarget SDK version 5.16.3.
+
+#### Version 5.16.2.0
+- Verified compatibility with myTarget SDK version 5.16.2.
+
+Built and tested with:
+- Google Mobile Ads SDK version 21.3.0.
+- MyTarget SDK version 5.16.2.
+
+#### Version 5.16.1.0
+- Verified compatibility with myTarget SDK version 5.16.1.
+
+Built and tested with:
+- Google Mobile Ads SDK version 21.3.0.
+- MyTarget SDK version 5.16.1.
+
 #### Version 5.16.0.0
 - Verified compatibility with myTarget SDK version 5.16.0.