+++ ThirdPartyAdapters/pangle/CHANGELOG.md
@@ -1,5 +1,27 @@
 ## Pangle Android Mediation Adapter Changelog
 
+#### Version 4.8.1.0.0 (In progress)
+- Verified compatibility with Pangle SDK version 4.8.1.0.
+- Updated the minimum required Google Mobile Ads SDK version to 21.4.0.
+
+Built and tested with:
+- Google Mobile Ads SDK version 21.4.0.
+- Pangle SDK version 4.8.1.0.
+
+#### Version 4.8.0.9.0
+- Verified compatibility with Pangle SDK version 4.8.0.9.
+
+Built and tested with:
+- Google Mobile Ads SDK version 21.3.0.
+- Pangle SDK version 4.8.0.9.
+
+#### Version 4.8.0.8.0
+- Verified compatibility with Pangle SDK version 4.8.0.8.
+
+Built and tested with:
+- Google Mobile Ads SDK version 21.3.0.
+- Pangle SDK version 4.8.0.8.
+
 #### Version 4.8.0.7.0
 - Verified compatibility with Pangle SDK version 4.8.0.7.